+++ netconfig/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class NetConfigConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'netconfig'
+    verbose_name = 'Network Configuration Backup'